requirements.txt
poetry
portia-sdk-python
openai